Update CodeQuality
Document `DeadlineParser`, `EventParser`, `ExitParser`, `FindCommandParser`, `MarkTaskParser`, `TaskListParser`, `TodoParser`, and `UnmarkTaskParser` with explanations of their functionality, parameters, and exceptions.

Improve documentation for better clarity and maintenance.
